Changing raw PID from UInt8 to UInt16.

+++ one_net/app/one_net_message.h
@@ -292,7 +292,7 @@ typedef struct
 //! calling functions.
 typedef struct
 {
-    UInt8 raw_pid; //! Raw PID of message
+    UInt16 raw_pid; //! Raw PID of message
     UInt16 msg_id; //! message id of this transaction(0 to 4095)
     UInt8 msg_type; //! message type (i.e. admin, app, etc.)
 } on_msg_hdr_t;
